Initial attempt to convert leaf-biomass-harvest to use new harvest libraries
git-svn-id: http://Marc-PC/svn/Full@3614 85f15e0e-5137-b144-a863-98bd542365a3

+++ trunk/leaf-biomass-harvest/branches/harvest-libs/src/EventsLog.cs
@@ -52,7 +52,7 @@ namespace Landis.Extension.LeafBiomassHarvest
         public int CohortsHarvestedComplete { set; get; }
 
         [DataFieldAttribute(Unit = FieldUnits.Count, Desc = "Species Cohorts Harvested by Species", SppList = true)]
-        public double[] CohortsHarvested_ { set; get; }
+        public int[] CohortsHarvested_ { set; get; }
 
         
         //[DataFieldAttribute(Unit = FiledUnits.None, Desc = "Initiation Row")]
